feat:Splitted deleteUserController form UsersController
