+++ src/application/controllers/UsersController.ts
@@ -1,29 +1,8 @@
 import UserNotFound from '@domain/exceptions/UserNotFound'
 import NotFoundException from '@application/exceptions/NotFoundException'
-import DeleteUserRequest from '@application/resources/DeleteUserRequest'
-import DeleteUserResponse from '@application/resources/DeleteUserResponse'
-import deleteUser from '@domain/use_cases/deleteUser'
 import UpdateUserRequest from '@application/resources/UpdateUserRequest'
 import UpdateUserResponse from '@application/resources/UpdateUserResponse'
 import updateUser from '@domain/use_cases/updateUser'
-
-const deleteUserController = async (
-  request: DeleteUserRequest
-): Promise<DeleteUserResponse> => {
-  try {
-    const userId = request.id
-
-    await deleteUser(userId)
-
-    return new DeleteUserResponse(userId)
-  } catch (e) {
-    if (e instanceof UserNotFound) {
-      throw new NotFoundException(e.message)
-    }
-
-    throw e
-  }
-}
 
 const updateUserController = async (
   request: UpdateUserRequest
@@ -43,4 +22,4 @@ const updateUserController = async (
   }
 }
 
-export { deleteUserController, updateUserController }
+export { updateUserController }
